requirement.txt
asgiref==3.3.1
chardet==4.0.0
Django==3.1.3
django-crispy-forms==1.11.0
idna==2.10
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
